First wip for updating to CLAP 0.18

+++ common/src/process.rs
@@ -3,9 +3,9 @@ use clap_sys::process::*;
 
 #[repr(i32)]
 pub enum ProcessStatus {
-    Continue = CLAP_PROCESS_CONTINUE,
-    ContinueIfNotQuiet = CLAP_PROCESS_CONTINUE_IF_NOT_QUIET,
-    Sleep = CLAP_PROCESS_SLEEP,
+    Continue = CLAP_PROCESS_CONTINUE as i32,
+    ContinueIfNotQuiet = CLAP_PROCESS_CONTINUE_IF_NOT_QUIET as i32,
+    Sleep = CLAP_PROCESS_SLEEP as i32,
 }
 
 #[repr(C)]
